+++ src/extension/debugConfigurationProvider.ts
@@ -6,7 +6,7 @@ import { TelemetryHelper } from "../common/telemetryHelper";
 import { Telemetry } from "../common/telemetry";
 import * as nls from "vscode-nls";
 import { PlatformType } from "./launchArgs";
-import { DirectDebugSession } from "../debugger/direct/directDebugSession";
+import { IWDPHelper } from "../debugger/direct/IWDPHelper";
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize = nls.loadMessageBundle();
 
@@ -94,7 +94,7 @@ export class ReactNativeDebugConfigProvider implements vscode.DebugConfiguration
             "type": DEBUG_TYPES.REACT_NATIVE_DIRECT,
             "request": "attach",
             "platform": PlatformType.iOS,
-            "port": DirectDebugSession.iOS_WEBKIT_DEBUG_PROXY_DEFAULT_PORT // 9221
+            "port": IWDPHelper.iOS_WEBKIT_DEBUG_PROXY_DEFAULT_PORT // 9221
         },
         "Debug Direct iOS (Preview) - Experimental": {
             "name": "Direct Debug iOS (Preview) - Experimental",
@@ -102,7 +102,7 @@ export class ReactNativeDebugConfigProvider implements vscode.DebugConfiguration
             "type": DEBUG_TYPES.REACT_NATIVE_DIRECT,
             "request": "launch",
             "platform": PlatformType.iOS,
-            "port": DirectDebugSession.iOS_WEBKIT_DEBUG_PROXY_DEFAULT_PORT // 9221
+            "port": IWDPHelper.iOS_WEBKIT_DEBUG_PROXY_DEFAULT_PORT // 9221
         },
         "Run Direct iOS (Preview) - Experimental": {
             "name": "Run Direct iOS (Preview) - Experimental",
